refactor: split builder.py in several files
