+++ src/anemoi/graphs/nodes/__init__.py
@@ -1,13 +1,13 @@
-from .builder import HEALPixNodes
-from .builder import HexNodes
-from .builder import LimitedAreaHEALPixNodes
-from .builder import LimitedAreaHexNodes
-from .builder import LimitedAreaNPZFileNodes
-from .builder import LimitedAreaTriNodes
-from .builder import LimitedAreaZarrDatasetNodes
-from .builder import NPZFileNodes
-from .builder import TriNodes
-from .builder import ZarrDatasetNodes
+from .builders.from_file import LimitedAreaNPZFileNodes
+from .builders.from_file import LimitedAreaZarrDatasetNodes
+from .builders.from_file import NPZFileNodes
+from .builders.from_file import ZarrDatasetNodes
+from .builders.from_healpix import HEALPixNodes
+from .builders.from_healpix import LimitedAreaHEALPixNodes
+from .builders.from_refined_icosahedron import HexNodes
+from .builders.from_refined_icosahedron import LimitedAreaHexNodes
+from .builders.from_refined_icosahedron import LimitedAreaTriNodes
+from .builders.from_refined_icosahedron import TriNodes
 
 __all__ = [
     "ZarrDatasetNodes",
